int id cambiado a long

+++ src/main/java/com/proyecto/services/BootcamperService.java
@@ -7,8 +7,8 @@ import com.proyecto.entities.Bootcamper;
 
 public interface BootcamperService {
     public List<Bootcamper> findAll(); //Busca todo
-    public Bootcamper findById(int idBootcamper); //Busca por Id
-    public void deleteById(int idBootcamper); //Borra
+    public Bootcamper findById(long idBootcamper); //Busca por Id
+    public void deleteById(long idBootcamper); //Borra
     public Bootcamper save (Bootcamper bootcamper); //Guarda y actualiza
 
 }
